# longaeva/docs/openapi.json
{
  "components": {
    "schemas": {
...
      "ForecastArchiveRequest": {
        "additionalProperties": false,
        "properties": {
          "kind": {
            "enum": [
              "retrospective",
              "prospective"
            ],
            "title": "Kind",
            "type": "string"
          }
        },
        "required": [
          "kind"
        ],
        "title": "ForecastArchiveRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ObservationReviewRead": {
        "properties": {
          "decisions": {
            "items": {
              "$ref": "#/components/schemas/ReviewDecisionRead"
            },
            "title": "Decisions",
            "type": "array"
          },
          "effective": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Effective"
          },
          "observation": {
            "$ref": "#/components/schemas/ObservationRead"
          }
        },
        "required": [
          "observation",
          "decisions"
        ],
        "title": "ObservationReviewRead",
        "type": "object"
      },
...
      "ReviewDecisionCreate": {
        "additionalProperties": false,
        "properties": {
          "corrected_payload": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Corrected Payload"
          },
          "decided_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decided At"
          },
          "decided_by": {
            "minLength": 1,
            "title": "Decided By",
            "type": "string"
          },
          "decision": {
            "enum": [
              "accept",
              "reject",
              "correct"
            ],
            "title": "Decision",
            "type": "string"
          },
          "observation_id": {
            "format": "uuid",
            "title": "Observation Id",
            "type": "string"
          },
          "rationale": {
            "minLength": 1,
            "title": "Rationale",
            "type": "string"
          }
        },
        "required": [
          "observation_id",
          "decision",
          "rationale",
          "decided_by"
        ],
        "title": "ReviewDecisionCreate",
        "type": "object"
      },
      "ReviewDecisionRead": {
        "properties": {
          "corrected_payload": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Corrected Payload"
          },
          "decided_at": {
            "format": "date-time",
            "title": "Decided At",
            "type": "string"
          },
          "decided_by": {
            "title": "Decided By",
            "type": "string"
          },
          "decision": {
            "title": "Decision",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "observation_id": {
            "format": "uuid",
            "title": "Observation Id",
            "type": "string"
          },
          "rationale": {
            "title": "Rationale",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "observation_id",
          "decision",
          "version",
          "rationale",
          "decided_at",
          "decided_by"
        ],
        "title": "ReviewDecisionRead",
        "type": "object"
      },
...
      "RunCreate": {
        "additionalProperties": false,
        "properties": {
          "cutoff_ts": {
            "format": "date-time",
            "title": "Cutoff Ts",
            "type": "string"
          },
          "n_paths": {
            "default": 5000,
            "maximum": 50000.0,
            "minimum": 1.0,
            "title": "N Paths",
            "type": "integer"
          },
          "n_quarters": {
            "default": 4,
            "maximum": 8.0,
            "minimum": 1.0,
            "title": "N Quarters",
            "type": "integer"
          },
          "scenario_id": {
            "format": "uuid",
            "title": "Scenario Id",
            "type": "string"
          },
          "seed": {
            "maximum": 2147483647.0,
            "minimum": 0.0,
            "title": "Seed",
            "type": "integer"
          },
          "switches": {
            "additionalProperties": {
              "type": "boolean"
            },
            "title": "Switches",
            "type": "object"
          }
        },
        "required": [
          "scenario_id",
          "cutoff_ts",
          "seed"
        ],
        "title": "RunCreate",
        "type": "object"
      },
      "RunRead": {
        "properties": {
          "baseline_run_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Baseline Run Id"
          },
          "code_version": {
            "title": "Code Version",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "cutoff_ts": {
            "format": "date-time",
            "title": "Cutoff Ts",
            "type": "string"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "finished_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "interventions": {
            "items": {},
            "title": "Interventions",
            "type": "array"
          },
          "interventions_hash": {
            "title": "Interventions Hash",
            "type": "string"
          },
          "job_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Job Id"
          },
          "lib_versions": {
            "additionalProperties": true,
            "title": "Lib Versions",
            "type": "object"
          },
          "n_paths": {
            "title": "N Paths",
            "type": "integer"
          },
          "n_quarters": {
            "title": "N Quarters",
            "type": "integer"
          },
          "origin_label": {
            "title": "Origin Label",
            "type": "string"
          },
          "outputs_hash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Outputs Hash"
          },
          "outputs_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Outputs Path"
          },
          "parameter_set_hash": {
            "title": "Parameter Set Hash",
            "type": "string"
          },
          "scenario_id": {
            "format": "uuid",
            "title": "Scenario Id",
            "type": "string"
          },
          "seed": {
            "title": "Seed",
            "type": "integer"
          },
          "source_manifest": {
            "items": {},
            "title": "Source Manifest",
            "type": "array"
          },
          "source_manifest_hash": {
            "title": "Source Manifest Hash",
            "type": "string"
          },
          "started_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "starting_state_hash": {
            "title": "Starting State Hash",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "switches": {
            "additionalProperties": true,
            "title": "Switches",
            "type": "object"
          }
        },
        "required": [
          "id",
          "scenario_id",
          "cutoff_ts",
          "origin_label",
          "n_quarters",
          "source_manifest",
          "source_manifest_hash",
          "parameter_set_hash",
          "starting_state_hash",
          "code_version",
          "seed",
          "n_paths",
          "switches",
          "interventions_hash",
          "lib_versions",
          "status",
          "created_at"
        ],
        "title": "RunRead",
        "type": "object"
      },
...
      "SavedDocumentRead": {
        "properties": {
          "document_links": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Document Links",
            "type": "object"
          },
          "filename": {
            "title": "Filename",
            "type": "string"
          },
          "key": {
            "title": "Key",
            "type": "string"
          },
          "markdown": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Markdown"
          },
          "owner_issue": {
            "title": "Owner Issue",
            "type": "string"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "status": {
            "enum": [
              "available",
              "pending"
            ],
            "title": "Status",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "key",
          "title",
          "filename",
          "status",
          "owner_issue"
        ],
        "title": "SavedDocumentRead",
        "type": "object"
      },
      "SavedReportRead": {
        "properties": {
          "ablations": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AblationReport"
              },
              {
                "type": "null"
              }
            ]
          },
          "document_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Document Key"
          },
          "extraction": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ExtractionReport"
              },
              {
                "type": "null"
              }
            ]
          },
          "forecast": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ForecastReport"
              },
              {
                "type": "null"
              }
            ]
          },
          "key": {
            "title": "Key",
            "type": "string"
          },
          "kind": {
            "enum": [
              "forecast",
              "ablations",
              "portfolio",
              "extraction",
              "prospective",
              "document",
              "pending"
            ],
            "title": "Kind",
            "type": "string"
          },
          "owner_issue": {
            "title": "Owner Issue",
            "type": "string"
          },
          "portfolio": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PortfolioReport"
              },
              {
                "type": "null"
              }
            ]
          },
          "prospective": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProspectiveReport"
              },
              {
                "type": "null"
              }
            ]
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "status": {
            "enum": [
              "available",
              "pending"
            ],
            "title": "Status",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "key",
          "title",
          "kind",
          "status",
          "owner_issue"
        ],
        "title": "SavedReportRead",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Standalone Visa business-simulation API (hackathon).",
    "title": "Longaeva",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/evaluation/documents/{document_key}": {
      "get": {
        "operationId": "get_saved_document_evaluation_documents__document_key__get",
        "parameters": [
          {
            "in": "path",
            "name": "document_key",
            "required": true,
            "schema": {
              "title": "Document Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SavedDocumentRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Saved Document",
        "tags": [
          "evaluation"
        ]
      }
    },
...
    "/evaluation/reports/{report_key}": {
      "get": {
        "operationId": "get_saved_report_evaluation_reports__report_key__get",
        "parameters": [
          {
            "in": "path",
            "name": "report_key",
            "required": true,
            "schema": {
              "title": "Report Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SavedReportRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Saved Report",
        "tags": [
          "evaluation"
        ]
      }
    },
...
    "/observations/{observation_id}/review": {
      "get": {
        "operationId": "get_observation_review_observations__observation_id__review_get",
        "parameters": [
          {
            "in": "path",
            "name": "observation_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Observation Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObservationReviewRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Observation Review",
        "tags": [
          "observations"
        ]
      }
    },
...
    "/review-decisions": {
      "get": {
        "operationId": "list_review_decisions_review_decisions_get",
        "parameters": [
          {
            "in": "query",
            "name": "observation_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Observation Id"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ReviewDecisionRead"
                  },
                  "title": "Response List Review Decisions Review Decisions Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Review Decisions",
        "tags": [
          "review"
        ]
      },
      "post": {
        "operationId": "create_review_decision_review_decisions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewDecisionCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewDecisionRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Review Decision",
        "tags": [
          "review"
        ]
      }
    },
    "/runs": {
      "get": {
        "operationId": "list_runs_runs_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RunRead"
                  },
                  "title": "Response List Runs Runs Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Runs",
        "tags": [
          "runs"
        ]
      },
      "post": {
        "operationId": "create_run_runs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Run",
        "tags": [
          "runs"
        ]
      }
    },
    "/runs/{run_id}": {
      "get": {
        "operationId": "get_run_runs__run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Run",
        "tags": [
          "runs"
        ]
      }
    },
    "/runs/{run_id}/forecasts": {
      "post": {
        "operationId": "archive_run_forecasts_runs__run_id__forecasts_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForecastArchiveRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ForecastRead"
                  },
                  "title": "Response Archive Run Forecasts Runs  Run Id  Forecasts Post",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Archive Run Forecasts",
        "tags": [
          "runs"
        ]
      }
    },
    "/runs/{run_id}/replay": {
      "post": {
        "operationId": "replay_saved_run_runs__run_id__replay_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReplayReport"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Replay Saved Run",
        "tags": [
          "runs"
        ]
      }
    },
    "/runs/{run_id}/results": {
      "get": {
        "operationId": "get_run_results_runs__run_id__results_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RunResultSummary"
                  },
                  "title": "Response Get Run Results Runs  Run Id  Results Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Run Results",
        "tags": [
          "runs"
        ]
      }
    },
...
  }
}